fix: move Editor into exported package
