+++ content/src/content/jcr_root/apps/core/email/components/commons/editor/dialog/segmenteditor/v1/segmenteditor/segmenteditor.html
@@ -13,7 +13,7 @@
   ~ See the License for the specific language governing permissions and
   ~ limitations under the License.
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/-->
-<coral-multifield data-sly-use.segmentEditor="com.adobe.cq.email.core.components.internal.commons.editor.dialog.segmenteditor.Editor"
+<coral-multifield data-sly-use.segmentEditor="com.adobe.cq.email.core.commons.editor.dialog.segmenteditor.Editor"
                   data-sly-use.iconTemplate="icon.html"
                   data-cmp-is="segmentEditor"
                   class="cmp-segmenteditor"
